+++ docs/about/release-notes.rst
@@ -13,6 +13,19 @@ To upgrade Sdklib to the latest version, use pip:
 Sdklib 1.8
 ==========
 
+Sdklib 1.8.7
+------------
+
+- Fix "The HTTP reason phrase should not be" behave step.
+- Add "The HTTP reason phrase should contain" behave step.
+- Add BaseHttpResponse class.
+
+Sdklib 1.8.6
+------------
+
+- Fix 11paths authorization bug.
+- Add insensitive (sort) parameter to to_key_val_list function.
+
 Sdklib 1.8.5.3
 --------------